Fix for a bunch of test issues.
