+++ news/2 Fixes/17285.md
@@ -1,0 +1,1 @@
+Improve detecting lines when using testing wrappers.
