+++ news/2 Fixes/17309.md
@@ -1,0 +1,1 @@
+Missing location info for `async def` functions.
